% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromOrthogroups.R
\name{fromOrthogroups}
\alias{fromOrthogroups}
\title{Read Orthologues results as a list}
\usage{
fromOrthogroups(orthogroups, annoGR_list, query_species = 1, split = ", ")
}
\arguments{
\item{orthogroups}{A dataframe with the gene ids for each orthologues group}

\item{annoGR_list}{A GRangesList object with the gene information. 
The names of each element should be contain all the gene ids saved in the 
orthogroups parameter. For each element, the 'gene_name' should be available
in the metadata.}

\item{query_species}{The species used as queryGR in \link{getGeneCluster}.}

\item{split}{character vector used to split the gene ids in the dataframe of 
orthogroups parameter.}
}
\value{
a list of queryGR and homologsList used by \link{getGeneCluster}.
}
\description{
Read Orthologues results as a list
}
\examples{
orthogroups <- read.delim(
     system.file('extdata/orthofinder/Orthogroups.tsv.gz',
                  package='geneClusterPattern'),
     row.names=1)
annoGR_list <- readRDS(
     system.file('extdata/orthofinder/grange.obj.rds',
                  package='geneClusterPattern'))
colnames(orthogroups) <- names(annoGR_list)
hres <- fromOrthogroups(orthogroups, annoGR_list)

}
