% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneClusterPattern.R
\docType{package}
\name{geneClusterPattern-package}
\alias{geneClusterPattern}
\alias{geneClusterPattern-package}
\title{Plot conserved gene pattern cross multiple species}
\description{
Understanding the conservation of gene neighborhoods across species provides important insights into genome organization, functional linkage, and evolutionary processes. This package is developed to facilitate the systematic characterization of orthologous gene arrangements surrounding a focal gene across multiple species.
Based on the orthology information, geneClusterPattern constructs an ordered representation of genes within a configurable genomic window upstream and downstream of the focal gene. The resulting ortholog gene pattern captures the relative arrangement of ortholog groups across species, enabling direct comparison of local genomic structure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jianhong/geneClusterPattern}
  \item Report bugs at \url{https://github.com/jianhong/geneClusterPattern/issues}
}

}
\author{
\strong{Maintainer}: Jianhong Ou \email{jou@morgridge.org} (\href{https://orcid.org/0000-0002-8652-2488}{ORCID})

Authors:
\itemize{
  \item Jianhong Ou \email{jou@morgridge.org} (\href{https://orcid.org/0000-0002-8652-2488}{ORCID})
  \item Kenneth Poss \email{kposs@morgridge.org} [funder]
}

}
