% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOrderScore.R
\name{geneOrderScore}
\alias{geneOrderScore}
\title{Calculate the gene order score}
\usage{
geneOrderScore(
  genesList,
  ids,
  ref,
  k = length(ids),
  max_gap = 1e+07,
  method = c("edit distance", "global alignment score", "spearman correlation",
    "kendall correlation", "non-random score", "pairs distance", "pairs direction"),
  output = c("score", "value matrix")
)
}
\arguments{
\item{genesList}{A list of GRanges.}

\item{ids}{IDs to be plotted. See \link{getGeneCluster}.}

\item{ref}{The reference species.}

\item{k}{The maximum number of nearest genes used in \link{getGeneCluster}.}

\item{max_gap}{The maximal gaps from the first ID in `ids`.}

\item{method}{The method to calculate the gene order score.
'edit distance', the gene order score is the mean of 
edit (Levenshtein) distance of given ids from different species.
Gene order score = k/length(ids)/(mean(adist)+1).
'global alignment score', the gene order score is the mean of global 
alignment score. see \link[pwalign]{pairwiseAlignment}.
'spearman correlation', the score is the mean of absolute value of
Spearman correlation of the genes appearance order.
'non-random score', the score is the mean of Jaccard index of 2-order and
3-order of ids for paired samples.
'pairs distance', the score is the mean of pairwise coordinates distance.
Score = k/length(ids)/(mean(abs(coordinates distance))+1).
'pairs direction', the score is the mean of alignment of
pairwise strand information.
The higher the gene order score, the higher the conservation of gene order.}

\item{output}{Output values}
}
\value{
the score or values matrix such as adist, or alignment scores.
}
\description{
Calculate the gene order score.
}
\examples{
fish <- readRDS(system.file('extdata', 'fish.rds',
package = 'geneClusterPattern'))
homologs <- readRDS(system.file('extdata', 'homologs.rds',
                                package = 'geneClusterPattern'))
queryGene <- 'inhbaa'
nearest10neighbors <- getGeneCluster(fish, queryGene, homologs, k=10)
genesList <- c(drerio=fish, homologs)[
c("hsapiens", "mmusculus", "drerio", "olatipes", "nfurzeri", "gaculeatus")]
# set zebrafish as reference
geneOrderScore(genesList, nearest10neighbors, ref='drerio')
# default first element is reference
geneOrderScore(genesList, nearest10neighbors, method='edit distance')
geneOrderScore(genesList, nearest10neighbors, method='global alignment score')
geneOrderScore(genesList, nearest10neighbors, method='spearman correlation')
geneOrderScore(genesList, nearest10neighbors, method='kendall correlation')
geneOrderScore(genesList, nearest10neighbors, method='non-random score')
geneOrderScore(genesList, nearest10neighbors, method='pairs distance')
geneOrderScore(genesList, nearest10neighbors, method='pairs direction')
}
