% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genePatternScore.R
\name{genePatternScore}
\alias{genePatternScore}
\title{Calculate the gene pattern score}
\usage{
genePatternScore(
  genesList,
  ids,
  k = length(ids),
  max_gap = 1e+07,
  output = c("score", "sd", "distance")
)
}
\arguments{
\item{genesList}{A list of GRanges.}

\item{ids}{IDs to be plotted. See \link{getGeneCluster}.}

\item{k}{The maximum number of nearest genes used in \link{getGeneCluster}.}

\item{max_gap}{The maximal gaps from the first ID in `ids`.}

\item{output}{Output values}
}
\value{
the standard deviation or edit distance values.
}
\description{
The gene pattern score is the normalized standard divination of 
edit (Levenshtein) distance of given ids from different species.
The maximal gene pattern score is 100.
Gene pattern score = 100 / ((sd(adist)+1)*k/length(ids))
}
\examples{
# example code

}
