% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneCluster.R
\name{getGeneCluster}
\alias{getGeneCluster}
\title{calculate the gene cluster}
\usage{
getGeneCluster(queryGR, queryGeneName, homologsList, k = 10, radius = 0)
}
\arguments{
\item{queryGR}{The query GRanges outputted by \link{grangesFromEnsemblIDs}.}

\item{queryGeneName}{The center gene used to search.}

\item{homologsList}{The homologs list outputted by \link{getHomologGeneList}.}

\item{k}{The maximum number of nearest genes.}

\item{radius}{The maximum number of gap genes for the search. If
radius is set to greater than 0, k parameter will be ignored.}
}
\value{
The ids of the cluster for the center gene.
}
\description{
A gene cluster is defined as the k nearest genes surrounding a reference gene
according to the rank matrix. The rank matrix is a gene-by-species matrix,
where each column contains the genomic ranks of genes for a given species.
}
\examples{
fish <- readRDS(system.file('extdata', 'fish.rds',
                            package = 'geneClusterPattern'))
homologs <- readRDS(system.file('extdata', 'homologs.rds',
                                package = 'geneClusterPattern'))
queryGene <- 'inhbaa'
nearestNeighbors <- getGeneCluster(fish, queryGene, homologs)

}
