% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneRank.R
\name{getGeneRank}
\alias{getGeneRank}
\title{Get the gene position orders}
\usage{
getGeneRank(genes)
}
\arguments{
\item{genes}{A GRanges object with names.}
}
\value{
A list with gene ranks named by the chromosome names.
}
\description{
Get the gene position orders without considering the strand info.
}
\examples{
# example code

}
