% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoFinder.R
\name{getHomologForOrthoFinder}
\alias{getHomologForOrthoFinder}
\title{Get the gene positions for output of OrthoFinder}
\usage{
getHomologForOrthoFinder(orthologs, mart)
}
\arguments{
\item{orthologs}{a two columns matrix}

\item{mart}{An Mart object}
}
\value{
A GRanges object with homolog_ensembl_gene_ids
}
\description{
Get the gene positions without considering the strand info.
}
\examples{
# example code

}
