% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHomologGeneList.R
\name{getHomologGeneList}
\alias{getHomologGeneList}
\title{Get the gene positions for homologs}
\usage{
getHomologGeneList(species, mart, ensembl_gene_ids, ...)
}
\arguments{
\item{species}{The target species for homologs query.}

\item{mart}{object of class Mart.}

\item{ensembl_gene_ids}{Ensembl gene ids.}

\item{...}{Parameters could be used by \link[biomaRt]{useEnsembl} except
`biomart`.}
}
\value{
A list of GRanges objects with gene positions.
}
\description{
Get the gene positions without considering the strand info.
}
\examples{
# example code

}
