% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneRankList.R
\name{getHomologGeneRankList}
\alias{getHomologGeneRankList}
\title{Get the gene position orders for homologs}
\usage{
getHomologGeneRankList(species, mart, ensembl_gene_ids, ...)
}
\arguments{
\item{species}{The target species for homologs query.}

\item{mart}{object of class Mart.}

\item{ensembl_gene_ids}{Ensembl gene ids.}

\item{...}{Parameters could be used by \link[biomaRt]{useEnsembl} except
`biomart`.}
}
\value{
A list of list with gene ranks named by the chromosome names.
}
\description{
Get the gene position orders without considering the strand info.
}
\examples{
# example code

}
