% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoFinder.R
\name{getHomologListForOrthoFinder}
\alias{getHomologListForOrthoFinder}
\title{Get the gene position orders for output of OrthoFinder}
\usage{
getHomologListForOrthoFinder(orthologs, marts)
}
\arguments{
\item{orthologs}{The output of \link{orthologPairsFromOrthoFinder}.}

\item{marts}{a list of object of class Mart.}
}
\value{
A list of list with gene ranks named by the chromosome names.
}
\description{
Get the gene position orders without considering the strand info.
}
\examples{
path <- system.file('extdata/orthofinder/Danio_rerio.GRCz11.pep.all.tsv.gz',
                  package='geneClusterPattern')
orthologs <- orthologPairsFromOrthoFinder(path)
annoGR_list <- readRDS(
     system.file('extdata/orthofinder/grange.obj.rds',
                  package='geneClusterPattern'))
homologs <- getHomologListForOrthoFinder(orthologs, annoGR_list[-1])
}
