% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grangesFromEnsemblIDs.R
\name{grangesFromEnsemblIDs}
\alias{grangesFromEnsemblIDs}
\title{create a GRanges object by given ensembl gene ids}
\usage{
grangesFromEnsemblIDs(mart, ensembl_gene_ids)
}
\arguments{
\item{mart}{object of class Mart.}

\item{ensembl_gene_ids}{Ensembl gene ids.}
}
\value{
An object of GRanges
}
\description{
Create a GRanges object by given ensembl gene ids
}
\examples{
if(interactive()){
  library(biomaRt)
  mart <- useMart('ensembl', 'drerio_gene_ensembl')
  grangesFromEnsemblIDs(mart, c('ENSDARG00000008803', 'ENSDARG00000017038'))
}

}
