% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessSpecies.R
\name{guessSpecies}
\alias{guessSpecies}
\title{Guess the species}
\usage{
guessSpecies(
  species,
  output = c("abbr", "scientific name", "mart", "taxid", "common name"),
  ...
)
}
\arguments{
\item{species}{A character(1L) to guess.}

\item{output}{Output type.}

\item{...}{Parameter will be passed to \link[biomaRt]{useEnsembl}.}
}
\value{
A character or Mart object.
}
\description{
Guess the species for a given string.
}
\examples{
if(interactive()){
  guessSpecies('zebrafish')
  guessSpecies('Japanese medaka')
  guessSpecies('stickleback')
}

}
