% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homologsFromEnsemblIDs.R
\name{homologsFromEnsemblIDs}
\alias{homologsFromEnsemblIDs}
\title{retrieve a homologs by given ensembl gene ids}
\usage{
homologsFromEnsemblIDs(species, mart, ensembl_gene_ids, ...)
}
\arguments{
\item{species}{The target species for homologs query.}

\item{mart}{object of class Mart.}

\item{ensembl_gene_ids}{Ensembl gene ids.}

\item{...}{Parameters could be used by \link[biomaRt]{useEnsembl} except
`biomart`.}
}
\value{
An object of GRanges
}
\description{
This function retrieves the user specified homologs for given ensembl gene
ids.
}
\examples{
if(interactive()){
  library(biomaRt)
  mart <- useMart('ensembl', 'drerio_gene_ensembl')
  homologsFromEnsemblIDs(
    'human',
    mart,
   c('ENSDARG00000008803', 'ENSDARG00000017038'))
}

}
