% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoFinder.R
\name{orthologPairsFromOrthoFinder}
\alias{orthologPairsFromOrthoFinder}
\title{Read orthoFinder Orthologues results as a list
Read orthoFinder Orthologues results as a list for each pair of orthologs.}
\usage{
orthologPairsFromOrthoFinder(path)
}
\arguments{
\item{path}{The path of the orthologues result file.}
}
\value{
a list of matrix with the paired orthologues.
}
\description{
Read orthoFinder Orthologues results as a list
Read orthoFinder Orthologues results as a list for each pair of orthologs.
}
\examples{
path <- system.file('extdata/orthofinder/Danio_rerio.GRCz11.pep.all.tsv.gz',
                  package='geneClusterPattern')
orthologs <- orthologPairsFromOrthoFinder(path)
}
