% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenePatterns.R
\name{plotGeneClusterPatterns}
\alias{plotGeneClusterPatterns}
\title{plot gene patters for given ids}
\usage{
plotGeneClusterPatterns(
  genesList,
  ids,
  additionalID,
  max_gap = 1e+07,
  colors,
  maxNonEssential = 5,
  alignType = c("local", "global"),
  label_size = 0.5
)
}
\arguments{
\item{genesList}{A list of GRanges.}

\item{ids}{IDs to be plotted. See \link{getGeneCluster}.}

\item{additionalID}{Other IDs should be included.}

\item{max_gap}{The maximal gaps from the first ID in `ids`.}

\item{colors}{The colors for genes}

\item{maxNonEssential}{The maximal number for non-essential genes between
two essential genes.}

\item{alignType}{Align the gene pattern by query gene ('local') or by the
clusters ('global').}

\item{The}{label cex value. default is 0.5.}
}
\value{
invisible list of plot data.
}
\description{
Plot the gene patters for given ids.
}
\examples{
fish <- readRDS(system.file('extdata', 'fish.rds',
                            package = 'geneClusterPattern'))
homologs <- readRDS(system.file('extdata', 'homologs.rds',
                                package = 'geneClusterPattern'))
queryGene <- 'inhbaa'
nearestNeighbors <- getGeneCluster(fish, queryGene, homologs)
genesList <- c(drerio=fish, homologs)
pgp <- plotGeneClusterPatterns(genesList, nearestNeighbors)
}
