% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruningSequences.R
\name{pruningSequences}
\alias{pruningSequences}
\title{Drop non-standard chromosomes}
\usage{
pruningSequences(genesList, negativePattern = "[_.M]")
}
\arguments{
\item{genesList}{A list of GRanges object}

\item{negativePattern}{The pattern to match the seqnames that does not need.}
}
\value{
A list of filtered GRanges
}
\description{
Drop non-standard chromosomes
}
\examples{
library(GenomicRanges)
gr <- GRanges(rep(c("chr2", "chr3", "chrM"), 2), IRanges(1:6, 10))
pruningSequences(list(gr))

}
