% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOrderScore.R
\name{geneOrderScore}
\alias{geneOrderScore}
\title{Calculate the gene order score}
\usage{
geneOrderScore(
  genesList,
  ids,
  ref,
  k = length(ids),
  max_gap = 1e+07,
  method = c("edit distance", "global alignment score", "spearman correlation"),
  output = c("score", "value matrix")
)
}
\arguments{
\item{genesList}{A list of GRanges.}

\item{ids}{IDs to be plotted. See \link{getGeneCluster}.}

\item{ref}{The reference species.}

\item{k}{The maximum number of nearest genes used in \link{getGeneCluster}.}

\item{max_gap}{The maximal gaps from the first ID in `ids`.}

\item{method}{The method to calculate the gene order score.
'edit distance', the gene order score is the mean of 
edit (Levenshtein) distance of given ids from different species.
Gene order score = mean(adist)*k/length(ids).
The higher the gene order score, the lower the conservation of the gene order.
'global alignment score', the gene order score is the mean of global 
alignment score. see \link[pwalign]{pairwiseAlignment}.
The higher the gene order score, the higher the conservation of the gene order.
'spearman correlation', the score is the mean of Spearman correlation of
the genes appearance order.}

\item{output}{Output values}
}
\value{
the score or values matrix such as adist, or alignment scores.
}
\description{
Calculate the gene order score.
}
\examples{
# example code

}
