% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOrderScore.R
\name{geneOrderScore}
\alias{geneOrderScore}
\title{Calculate the gene order score}
\usage{
geneOrderScore(
  genesList,
  ids,
  ref,
  k = length(ids),
  max_gap = 1e+07,
  output = c("score", "distance")
)
}
\arguments{
\item{genesList}{A list of GRanges.}

\item{ids}{IDs to be plotted. See \link{getGeneCluster}.}

\item{ref}{The reference species.}

\item{k}{The maximum number of nearest genes used in \link{getGeneCluster}.}

\item{max_gap}{The maximal gaps from the first ID in `ids`.}

\item{output}{Output values}
}
\value{
the standard deviation or edit distance values.
}
\description{
The gene order score is the mean of 
edit (Levenshtein) distance of given ids from different species.
Gene order score = mean(adist)*k/length(ids).
The higher the gene order score, the lower the conservation of the gene order
}
\examples{
# example code

}
