% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneCluster.R
\name{getGeneCluster}
\alias{getGeneCluster}
\title{calculate the gene cluster}
\usage{
getGeneCluster(queryGR, queryGeneName, homologsList, k = 10, radius = 0)
}
\arguments{
\item{queryGR}{The query GRanges outputed by \link{grangesFromEnsemblIDs}.}

\item{queryGeneName}{The center gene used to search.}

\item{homologsList}{The homologs list outputed by \link{getHomologGeneList}.}

\item{k}{The maximum number of nearest genes.}

\item{radius}{The maximum number of gap genes for the search. If
radius is set to greater than 0, k parameter will be ignored.}
}
\value{
The ids of the cluster for the center gene.
}
\description{
A short description...
}
\examples{
# example code

}
