% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoFinder.R
\name{orthologPairsFromOrthoFinder}
\alias{orthologPairsFromOrthoFinder}
\title{Read orthoFinder Orthologues results as a list}
\usage{
orthologPairsFromOrthoFinder(path)
}
\arguments{
\item{path}{The path of the orthologues result file eg:
'primary_transcripts/OrthoFinder/Results_Nov01/Orthologues/Danio_rerio.GRCz11.pep.all.tsv'}
}
\value{
a list of matrix with the paired orthologues.
}
\description{
Read orthoFinder Orthologues results as a list
}
